batch_layer/requirements.txt
kafka-python
hdfs
pyspark
numpy==1.21.5
# scikit-learn
# xgboost
mysql-connector-python